+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## 1.1.1 under development
 
-- no changes in this release.
+- Enh #48 Throw an `InvalidArgumentException` if the retry delay is less than one millisecond (devanych)
 
 ## 1.1.0 October 19, 2021